requirements.txt
Django==5.2
djangorestframework==3.15.2
setuptools==69.0.0
web3==6.11.3
eth-account==0.10.0
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
hdwallet==2.2.1
mnemonic==0.20
django-cors-headers==4.3.0
